# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOC Analyzer",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Create Upload",
        "operationId": "create_upload_api_uploads_post",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_upload_api_uploads_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "List Uploads",
        "description": "Newest first. Summary fields only (no bad-line samples).",
        "operationId": "list_uploads_api_uploads_get",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UploadOut"
                  },
                  "title": "Response List Uploads Api Uploads Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/uploads/{upload_id}": {
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "Get Upload",
        "description": "One upload with status, progress and parse results. The status page polls this.",
        "operationId": "get_upload_api_uploads__upload_id__get",
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Upload Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "uploads"
        ],
        "summary": "Delete Upload",
        "operationId": "delete_upload_api_uploads__upload_id__delete",
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Upload Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BadLineOut": {
        "properties": {
          "line_no": {
            "type": "integer",
            "title": "Line No"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "raw": {
            "type": "string",
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "line_no",
          "reason",
          "raw"
        ],
        "title": "BadLineOut"
      },
...
      "Body_create_upload_api_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "log_timezone": {
            "type": "string",
            "title": "Log Timezone",
            "default": "UTC"
          },
          "format": {
            "type": "string",
            "enum": [
              "auto",
              "csv",
              "json"
            ],
            "title": "Format",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_upload_api_uploads_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadCreated": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "log_timezone": {
            "type": "string",
            "title": "Log Timezone"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "line_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Count"
          },
          "bad_line_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bad Line Count"
          },
          "previous_uploads": {
            "type": "integer",
            "title": "Previous Uploads"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "format",
          "log_timezone",
          "status",
          "progress",
          "size_bytes",
          "sha256",
          "uploaded_by",
          "created_at",
          "error",
          "line_count",
          "bad_line_count",
          "previous_uploads"
        ],
        "title": "UploadCreated"
      },
      "UploadDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "log_timezone": {
            "type": "string",
            "title": "Log Timezone"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "line_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Count"
          },
          "bad_line_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bad Line Count"
          },
          "bad_line_samples": {
            "items": {
              "$ref": "#/components/schemas/BadLineOut"
            },
            "type": "array",
            "title": "Bad Line Samples"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "disabled_detectors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Disabled Detectors"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "format",
          "log_timezone",
          "status",
          "progress",
          "size_bytes",
          "sha256",
          "uploaded_by",
          "created_at",
          "error",
          "line_count",
          "bad_line_count",
          "bad_line_samples",
          "completed_at",
          "disabled_detectors"
        ],
        "title": "UploadDetail",
        "description": "UploadOut plus the larger fields, only for a single upload."
      },
      "UploadOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "log_timezone": {
            "type": "string",
            "title": "Log Timezone"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "line_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Count"
          },
          "bad_line_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bad Line Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "format",
          "log_timezone",
          "status",
          "progress",
          "size_bytes",
          "sha256",
          "uploaded_by",
          "created_at",
          "error",
          "line_count",
          "bad_line_count"
        ],
        "title": "UploadOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}